Windows port: experimental ccl: MSVC narrowing casts and pimpl-completeness fixes
- all_gather_async, all_gather_concat_heads_fused, all_gather_minimal_matmul_async,
  all_reduce_async, all_to_all_async(_generic), all_to_all_dispatch_metadata,
  deepseek_moe_reduce_scatter, llama_all_gather_matmul_async, llama_reduce_scatter(_create_heads),
  llama_reduce_scatter_matmul, minimal_matmul_strided_reduce_scatter_async, moe_compute,
  neighbor_pad_async, neighbor_pad_halo, reduce_scatter_minimal_async(_direct),
  ring_attention_all_gather_async, send_recv_async (recv/send, direct and indirect),
  slice_reshard_async, strided_all_gather_async, strided_reduce_scatter_async:
  add static_cast<uint32_t>(...) around DeviceAddr/size_t/int expressions
  (.buffer()->address(), .address(), .size(), CoreCoord.x/.y, etc.) that MSVC treats
  as narrowing inside std::vector<uint32_t> and other braced-init-lists used for
  compile-time and runtime kernel args.
- ring_attention_all_gather_async: fix two CBDescriptor aggregate-init failures
  (C2440) caused by size_t packet-header-size expressions narrowing into
  uint32_t total_size/page_size fields.
- all_gather_async, all_to_all_dispatch_metadata, llama_reduce_scatter(_create_heads),
  minimal_matmul_strided_reduce_scatter_async, moe/selective_reduce_combine,
  reduce_scatter_minimal_async, ring_attention_all_gather_async, strided_all_gather_async,
  strided_reduce_scatter_async, dit_fused_distributed_groupnorm/rmsnorm,
  all_gather_minimal_matmul_async: include tt_metal/impl/buffers/global_semaphore_impl.hpp
  wherever operation_attributes_t (or a struct it nests) stores a GlobalSemaphore by
  value (in std::optional or std::vector). MSVC's aggregate/attribute reflection
  machinery in device_operation.hpp and tt_stl/reflection.hpp needs the
  GlobalSemaphoreImpl pimpl type to be complete at these points; GCC/Clang do not.

+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_gather_async/device/all_gather_via_broadcast_factory.cpp
@@ -228,18 +228,19 @@ tt::tt_metal::ProgramDescriptor build_descriptor_at(
         // (semaphores live on workload_descriptor.semaphores), so they may
         // remain embedded plain uint32 values.
         std::vector<uint32_t> writer_rt_args = {
-            output_tensor.buffer()->address(),  // tensor_address0  (replaced with Buffer* below)
-            semaphore.address(),                // out_ready_sem_bank_addr (absolute address)
-            barrier_semaphore.address(),        // barrier_sem
+            static_cast<uint32_t>(output_tensor.buffer()->address()),  // tensor_address0  (replaced with Buffer*
+                                                                        // below)
+            static_cast<uint32_t>(semaphore.address()),         // out_ready_sem_bank_addr (absolute address)
+            static_cast<uint32_t>(barrier_semaphore.address()),  // barrier_sem
             output_tile_id_start,
             output_tile_id_end,
-            wait_output_semaphore,     // wait_output_semaphore
-            reset_global_semaphore,    // reset_global_semaphore
-            drain_sync_core.x,         // out_ready_sem_noc0_x
-            drain_sync_core.y,         // out_ready_sem_noc0_y
-            out_ready_sem_wait_value,  // out_ready_sem_wait_value
-            barrier_core.x,            // barrier_sem_noc0_x
-            barrier_core.y             // barrier_sem_noc0_y
+            wait_output_semaphore,                    // wait_output_semaphore
+            reset_global_semaphore,                    // reset_global_semaphore
+            static_cast<uint32_t>(drain_sync_core.x),  // out_ready_sem_noc0_x
+            static_cast<uint32_t>(drain_sync_core.y),  // out_ready_sem_noc0_y
+            out_ready_sem_wait_value,                  // out_ready_sem_wait_value
+            static_cast<uint32_t>(barrier_core.x),     // barrier_sem_noc0_x
+            static_cast<uint32_t>(barrier_core.y)      // barrier_sem_noc0_y
         };
         auto num_connections = (int)forward_coord.has_value() + (int)backward_coord.has_value();
         writer_rt_args.push_back(num_connections);
